+++ src/vs/workbench/contrib/chat/browser/chat.contribution.ts
@@ -80,7 +80,8 @@ import { ChatPasteProvidersFeature } from './chatPasteProviders.js';
 import { QuickChatService } from './chatQuick.js';
 import { ChatQuotasService, IChatQuotasService } from '../common/chatQuotasService.js';
 import { ChatResponseAccessibleView } from './chatResponseAccessibleView.js';
-import { ChatEntitlementsService, ChatSetupContribution, IChatEntitlementsService } from './chatSetup.js';
+import { ChatEntitlementsService, ChatSetupContribution } from './chatSetup.js';
+import { IChatEntitlementsService } from '../common/chatEntitlementsService.js';
 import { ChatVariablesService } from './chatVariables.js';
 import { ChatWidgetService } from './chatWidget.js';
 import { ChatCodeBlockContextProviderService } from './codeBlockContextProviderService.js';
